+++ core/primitives-core/src/runtime/fees.rs
@@ -83,9 +83,9 @@ impl RuntimeFeesConfig {
             pessimistic_gas_price_inflation_ratio: Rational32::new(103, 100),
             action_fees: enum_map::enum_map! {
                 ActionCosts::create_account => Fee {
-                    send_sir: 99607375000,
-                    send_not_sir: 99607375000,
-                    execution: 99607375000,
+                    send_sir: 3_850_000_000_000,
+                    send_not_sir: 3_850_000_000_000,
+                    execution: 3_850_000_000_000,
                 },
                 ActionCosts::delete_account => Fee {
                     send_sir: 147489000000,
@@ -157,11 +157,10 @@ impl RuntimeFeesConfig {
                     send_not_sir: 59357464,
                     execution: 59357464,
                 },
-                #[cfg(feature = "protocol_feature_nep366_delegate_action")]
                 ActionCosts::delegate => Fee {
-                    send_sir: 2319861500000,
-                    send_not_sir: 2319861500000,
-                    execution: 2319861500000,
+                    send_sir: 200_000_000_000,
+                    send_not_sir: 200_000_000_000,
+                    execution: 200_000_000_000,
                 },
             },
         }