+++ integration-tests/src/tests/network/peer_handshake.rs
@@ -23,6 +23,8 @@ fn make_peer_manager(
     boot_nodes: Vec<(&str, std::net::SocketAddr)>,
     peer_max_count: u32,
 ) -> actix::Addr<PeerManagerActor> {
+    use near_async::messaging::Sender;
+
     let mut config = config::NetworkConfig::from_seed(seed, node_addr);
     config.peer_store.boot_nodes = convert_boot_nodes(boot_nodes);
     config.max_num_peers = peer_max_count;
@@ -34,7 +36,7 @@ fn make_peer_manager(
         near_store::db::TestDB::new(),
         config,
         Arc::new(near_network::client::Noop),
-        Arc::new(near_network::client::Noop),
+        Sender::noop(),
         GenesisId::default(),
     )
     .unwrap()